refactor: separate file for account page container componenrt
